neurovoice-backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
librosa==0.10.0
numpy==1.24.3
scipy==1.11.2
soundfile==0.12.1
